Restructure mcp_server package layout
Moved all mcp_server files from src/backend/v3/mcp_server/ to src/mcp_server/, updating imports and package structure accordingly. Added src/mcp_server/__init__.py to define the package and version. Updated pyproject.toml script entry point. Also removed a redundant utils/__init__.py file.
