+++ src/mcp_server/Dockerfile
