+++ src/mcp_server/README_NEW.md
