+++ src/mcp_server/__init__.py
@@ -1,0 +1,3 @@
+"""MACAE MCP Server package."""
+
+__version__ = "0.1.0"
