+++ src/mcp_server/docker-compose.yml
