+++ src/mcp_server/services/general_service.py
