+++ src/mcp_server/utils/formatters.py
